+++ mobile/android/android-components/components/concept/engine/src/main/java/mozilla/components/concept/engine/EngineSession.kt
@@ -7,6 +7,7 @@ package mozilla.components.concept.engine
 import android.graphics.Bitmap
 import androidx.annotation.CallSuper
 import mozilla.components.concept.engine.media.Media
+import mozilla.components.concept.engine.media.RecordingDevice
 import mozilla.components.concept.engine.permission.PermissionRequest
 import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.engine.window.WindowRequest
@@ -49,6 +50,7 @@ abstract class EngineSession(
         fun onMediaAdded(media: Media) = Unit
         fun onMediaRemoved(media: Media) = Unit
         fun onCrashStateChange(crashed: Boolean) = Unit
+        fun onRecordingStateChanged(devices: List<RecordingDevice>) = Unit
 
         /**
          * The engine received a request to load a request.